fix: add Provides to chromium
The issue is that when constructing the O6/06N OS image, the upstream
Chromium package may be installed rather than the expected CIX Chromium
from cix-prebuilt[0]. The CIX Chromium package has the same name as the
Debian upstream one. The Debian config file was expected to prioritise
the CIX Chromium[1]. However, the issue still arises when Chromium was
installed before `cix-profiles`.

To solve the issue, a `Provides` object `chromium-cix` was added to the
CIX Chromium package, so as long as we use 'chromium-cix' to install
Chromium, the CIX Chromium will be installed.

[0]: https://github.com/radxa-pkg/cix-prebuilt/
[1]: https://github.com/radxa-pkg/cix-profiles/blob/87d2547a0c7067547bdbaa83a072738cc80d542f/apt/radxa-cd8180#L1-L7

Signed-off-by: Zhaoming Luo <luozhaoming@radxa.com>